refactor(cli): split autofix.ts into pipeline + slim wrapper (Problem 2)
Splits packages/cli/src/commands/autofix.ts (1991 lines) into:
- packages/cli/src/autofix-pipeline.ts (1840 lines, NEW) — runAutofix
  body + AutofixArgs + AutofixDeps + Exec types + Node default runners
  + parseVerdictFile + helpers (finalizeIteration / bailResult /
  remainingBlockersFrom / buildWorker)
- packages/cli/src/commands/autofix.ts (188 lines) — HELP + parseArgv
  + autofix(argv) CLI entry + renderAutofixSummary + pass-through
  re-exports so existing test imports stay green

Why: Problem 3 (central-plane SaaS endpoints) needs to import the
pipeline as a library. Pre-split, runAutofix lived in a file dominated
by CLI ergonomics. After this refactor, central-plane can:

  import { runAutofix } from "@conclave-ai/cli/dist/autofix-pipeline.js";
  await runAutofix(args, { git: customGit, gh: customGh, ... });

Pipeline still imports node:child_process / node:fs / node:util at
module level (used by default* runners). Central-plane injects
overrides to skip those when on bundlers that can't handle node:*.
That readiness is Problem 3 territory; this PR is structural separation.

Behavior unchanged. All comments preserved verbatim. All public
exports retained. --prd flag (Phase 3) preserved in both AutofixArgs
+ parseArgv. 47/47 task suites pass.
